requirements.txt
python-docx
docx2pdf
dash
dash_bootstrap_components
pillow
transformers
torch
tqdm